+++ src/test/shared/Vector2fProductTest.java
@@ -1,4 +1,4 @@
-package Test;
+package test.shared;
 
 import static org.junit.Assert.*;
